Refactor: Warnings are emitted from the Analyzer instead of the Task
