--- src/Analyzer/AnalyzerReleases.Shipped.md
@@ -1,10 +1,0 @@
-; Shipped analyzer releases
-; https://github.com/dotnet/roslyn-analyzers/blob/master/src/Microsoft.CodeAnalysis.Analyzers/ReleaseTrackingAnalyzers.Help.md
-
-## Release 3.1.0
-
-### New Rules
-
-Rule ID | Category | Severity | Notes
---------|----------|----------|--------------------
-DOC001  | Documentation |  Warning | See https://github.com/dotnet/roslyn/issues/66188
