--- src/Analyzer/AnalyzerReleases.Unshipped.md
@@ -1,2 +1,0 @@
-; Unshipped analyzer release
-; https://github.com/dotnet/roslyn-analyzers/blob/master/src/Microsoft.CodeAnalysis.Analyzers/ReleaseTrackingAnalyzers.Help.md
