+++ src/Shared/DeclaredReferences.cs
@@ -1,0 +1,7 @@
+﻿namespace ReferenceTrimmer.Shared;
+
+internal record DeclaredReferences(IReadOnlyList<DeclaredReference> References);
+
+internal record DeclaredReference(string AssemblyName, DeclaredReferenceKind Kind, string Spec);
+
+internal enum DeclaredReferenceKind { Reference, ProjectReference, PackageReference }
