+++ src/Shared/NetStandardPolyfills.cs
@@ -1,0 +1,3 @@
+﻿namespace System.Runtime.CompilerServices;
+
+internal sealed class IsExternalInit { }
