--- src/Tasks/NetFrameworkPolyfills.cs
@@ -1,4 +1,0 @@
-﻿namespace System.Runtime.CompilerServices
-{
-    internal sealed class IsExternalInit { }
-}
