+++ src/Tasks/ReferenceTrimmer.Tasks.csproj
@@ -5,5 +5,9 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.Build.Utilities.Core" Version="17.3.1" />
     <PackageReference Include="NuGet.ProjectModel" Version="6.3.0" />
+    <PackageReference Include="System.Text.Json" Version="7.0.2" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="..\Shared\*.cs" />
   </ItemGroup>
 </Project>
